requirements.txt
Django==4.1.7
django-environ==0.9.0
django_debug_toolbar==3.8.1
psycopg2-binary==2.9.5
